rename build.task to build-application.task
