+++ scopes/harmony/application/application.main.runtime.ts
@@ -15,7 +15,7 @@ import { DeploymentProvider } from './deployment-provider';
 import { AppNotFound } from './exceptions';
 import { ApplicationAspect } from './application.aspect';
 import { AppListCmdDeprecated } from './app-list.cmd';
-import { AppsBuildTask } from './build.task';
+import { AppsBuildTask } from './build-application.task';
 import { RunCmd } from './run.cmd';
 import { AppService } from './application.service';
 import { AppCmd, AppListCmd } from './app.cmd';
