+++ scopes/harmony/application/index.ts
@@ -7,6 +7,6 @@ export { ApplicationType } from './application-type';
 export { AppDeployContext } from './app-deploy-context';
 export { AppBuildContext } from './app-build-context';
 export { AppBuildResult } from './app-build-result';
-export { ARTIFACTS_DIR_NAME as APPS_ARTIFACTS_DIR_NAME } from './build.task';
+export { ARTIFACTS_DIR_NAME as APPS_ARTIFACTS_DIR_NAME } from './build-application.task';
 export type { AppsEnv } from './apps-env-type';
 export { AppTypeList } from './app-type-list';
